Shuffle the deck when it is initialized instead of choosing a random card every draw or reveal. Those actions now just pop the top card from the deck.
The motivation for this was for testing a while game - I wanted to get a new game but then replace
the deck with a known array so that the rest of the game actions would be deterministic. This was much easier
to do when the 'draw' and 'reveal' actions are known to just pop the top card and not involve any random numbers.
So now I can set up a test game state after all the randomization has already happened.

Used the Fisher–Yates algorithm and did a bunch of testing to make sure I had it correct and to see if I could
get it to "shuffle" into the original state. Because if it did my test will fail since I'm asserting that the
shuffled deck is not the same as the original one. I satisfied myself that the odds were low enough that it
wouldn't be a problem.